# swagger2.0.yml
swagger: "2.0"
info:
  description: "This is a sample of a creating a board"
  version: "1.0.0-beta"
  title: "Swagger Todo BoardController"
  license:
    name: "Apache 2.0"
    url: "http://localhost:8080"
host: "localhost"
...
schemes:
- "https"
paths:
...
  /user/login:
    get:
      tags:
      - "users"
      summary: "Logs user into the system"
      description: ""
      operationId: "loginUser"
      produces:
      - "application/json"
      parameters:
      - name: "email"
        in: "query"
        description: "The user name for login"
        required: true
        type: "string"
      - name: "password"
        in: "query"
        description: "The password for login in clear text"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid username/password supplied"
...